WIP: Store all cops in a Set for faster checks
When profiling the startup time of `standard` (by running `standard
--help`), I noticed that quite a lot of time was spent calling the
`RuboCop::Cop::Base#match?` method (743810 times).

Looking at the logic for that method I thought that we could store
the names we are interested in matching against (`badge.department.name` and
`cop_name`) when we add the cop to the registry. This way we don't have
iterate through all the cops and look for a match, but instead can do a
quick lookup in our `@cops_set` to see if we include it or not.

This seems to shave off around ~0.4s seconds when running the `standard
--help` command on my machine.

After:
```
❯ time bin/run --help >/dev/null
bin/run --help > /dev/null  1.15s user 0.32s system 96% cpu 1.522 total
```

After:
```
❯ time bin/run --help >/dev/null
bin/run --help > /dev/null  0.86s user 0.27s system 95% cpu 1.179 total
```